removed unnecessary casts

+++ app/controllers/importutilities/comparators/StartNrComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class StartNrComparator implements Comparator<RiderStageConnection> {
     @Override
     public int compare(RiderStageConnection a, RiderStageConnection b) {
-        return (int)(a.getRider().getStartNr() - b.getRider().getStartNr());
+        return (a.getRider().getStartNr() - b.getRider().getStartNr());
     }
 }
